feat: Add local CI validation with ACT and fix test failures
- Set up ACT configuration for running GitHub Actions locally
- Add comprehensive local CI guide documentation
- Fix 5 test failures in test_decorator_edge_cases.py (mocking issues)
- Update test mocking paths for DataProtectionEngine
- Clean up test files and add proper fixtures
- Add validation for all CI workflows (test, security, build-test)
- Improve pre-commit hook compatibility

Resolves local CI validation requirements for faster development feedback
