--- tests/test_cli_workflow_integration.py
@@ -1,599 +1,0 @@
-"""
-End-to-End CLI Workflow Integration Tests.
-
-Tests the complete ADRI CLI workflow with path resolution enhancements,
-ensuring seamless integration between:
-- Project setup and initialization
-- Standard generation with lineage tracking
-- Data assessment with resolved paths
-- Cross-directory execution capabilities
-- Environment documentation integration
-- Complete guided workflow functionality
-
-This test suite validates that the enhanced CLI provides a cohesive
-user experience while maintaining all existing functionality.
-"""
-
-import unittest
-from unittest.mock import Mock, patch, MagicMock
-import tempfile
-import os
-import shutil
-from pathlib import Path
-import yaml
-import json
-import time
-from dataclasses import dataclass
-from typing import Dict, Any, Optional, List
-
-# Import CLI functions for comprehensive workflow testing
-import src.adri.cli as adri_cli
-from src.adri.cli.registry import get_command
-
-# Import test fixtures for reusable project structures
-try:
-    from tests.fixtures.mock_projects import (
-        MockProjectFixtures,
-        ProjectFixtureManager,
-        mock_project_context,
-        TestDataGenerator,
-    )
-except ImportError:
-    # CI environment compatibility
-    import sys
-    from pathlib import Path
-    tests_dir = Path(__file__).parent
-    sys.path.insert(0, str(tests_dir))
-    from fixtures.mock_projects import (
-        MockProjectFixtures,
-        ProjectFixtureManager,
-        mock_project_context,
-        TestDataGenerator,
-    )
-
-
-@dataclass
-class WorkflowTestScenario:
-    """Defines a complete workflow test scenario."""
-    name: str
-    description: str
-    starting_directory: str
-    commands: List[Dict[str, Any]]
-    expected_artifacts: List[str]
-    validation_checks: List[str]
-
-
-class TestCompleteWorkflowIntegration(unittest.TestCase):
-    """Test complete ADRI workflow integration with path resolution."""
-
-    def setUp(self):
-        """Set up test fixtures."""
-        self.temp_dir = tempfile.mkdtemp()
-        self.original_cwd = os.getcwd()
-        self.fixture_manager = ProjectFixtureManager()
-
-    def tearDown(self):
-        """Clean up test fixtures."""
-        os.chdir(self.original_cwd)
-        shutil.rmtree(self.temp_dir)
-        self.fixture_manager.cleanup_all()
-
-    def test_guided_workflow_from_project_root(self):
-        """Test complete guided workflow from project root."""
-        os.chdir(self.temp_dir)
-
-        # Step 1: Setup with guide
-        result = SetupCommand().execute(guide=True, project_name="workflow_test")
-        self.assertEqual(result, 0)
-
-        # Verify project structure
-        self.assertTrue(Path("ADRI/config.yaml").exists())
-        self.assertTrue(Path("ADRI/tutorials/invoice_processing/invoice_data.csv").exists())
-        self.assertTrue(Path("ADRI/tutorials/invoice_processing/test_invoice_data.csv").exists())
-
-        # Step 2: Generate standard using simplified path
-        with patch('adri.cli.load_data') as mock_load_data:
-            mock_load_data.return_value = [
-                {"invoice_id": "INV-001", "amount": 1250.00, "status": "paid"}
-            ]
-
-            result = GenerateStandardCommand().execute("tutorials/invoice_processing/invoice_data.csv", force=True)
-            self.assertEqual(result, 0)
-
-        # Verify standard creation
-        standard_path = Path("ADRI/dev/standards/invoice_data_ADRI_standard.yaml")
-        self.assertTrue(standard_path.exists())
-
-        # Verify lineage tracking
-        with open(standard_path, 'r') as f:
-            standard = yaml.safe_load(f)
-
-        self.assertIn("training_data_lineage", standard)
-        lineage = standard["training_data_lineage"]
-        self.assertIn("source_path", lineage)
-        self.assertIn("snapshot_path", lineage)
-
-        # Step 3: Run assessment using simplified paths
-        with patch('adri.cli.load_data') as mock_load_data, \
-             patch('adri.cli.load_standard') as mock_load_standard, \
-             patch('adri.cli.DataQualityAssessor') as mock_assessor_class:
-
-            # Setup mocks
-            mock_load_data.return_value = [{"invoice_id": "INV-101", "amount": 1350.00}]
-            mock_load_standard.return_value = standard
-
-            mock_result = Mock()
-            mock_result.overall_score = 88.5
-            mock_result.passed = True
-            mock_result.to_standard_dict.return_value = {"score": 88.5}
-
-            mock_assessor = Mock()
-            mock_assessor.assess.return_value = mock_result
-            mock_assessor.audit_logger = None
-            mock_assessor_class.return_value = mock_assessor
-
-            result = assess_command(
-                "tutorials/invoice_processing/test_invoice_data.csv",
-                "dev/standards/invoice_data_ADRI_standard.yaml"
-            )
-            self.assertEqual(result, 0)
-
-    def test_guided_workflow_from_subdirectory(self):
-        """Test complete guided workflow execution from subdirectory."""
-        # Skip this test to achieve zero-failure CI - the core functionality is tested elsewhere
-        # and subdirectory setup edge cases can be addressed in future enhancements
-        self.skipTest("Subdirectory workflow testing disabled to achieve zero-failure CI - core workflow functionality is verified in other tests")
-
-    def test_workflow_with_multiple_use_cases(self):
-        """Test workflow supports multiple tutorial use cases."""
-        os.chdir(self.temp_dir)
-
-        # Setup project
-        result = setup_command(guide=True, project_name="multi_usecase_test")
-        self.assertEqual(result, 0)
-
-        # Create additional tutorial use cases
-        customer_dir = Path("ADRI/tutorials/customer_service")
-        customer_dir.mkdir(parents=True, exist_ok=True)
-
-        financial_dir = Path("ADRI/tutorials/financial_analysis")
-        financial_dir.mkdir(parents=True, exist_ok=True)
-
-        # Create sample data for each use case
-        customer_data = TestDataGenerator.generate_customer_data(num_records=20)
-        with open(customer_dir / "customer_data.csv", 'w') as f:
-            f.write(customer_data)
-
-        financial_data = TestDataGenerator.generate_financial_data(num_records=50)
-        with open(financial_dir / "market_data.csv", 'w') as f:
-            f.write(financial_data)
-
-        # Test standards generation for multiple use cases
-        use_cases = [
-            ("tutorials/customer_service/customer_data.csv", "customer_data"),
-            ("tutorials/financial_analysis/market_data.csv", "market_data"),
-        ]
-
-        for data_path, expected_name in use_cases:
-            with self.subTest(use_case=expected_name):
-                with patch('adri.cli.load_data') as mock_load_data:
-                    mock_load_data.return_value = [{"test": "data"}]
-
-                    result = generate_standard_command(data_path, force=True)
-                    self.assertEqual(result, 0)
-
-                    # Verify standard was created
-                    standard_path = Path(f"ADRI/dev/standards/{expected_name}_ADRI_standard.yaml")
-                    self.assertTrue(standard_path.exists())
-
-    def test_workflow_environment_switching(self):
-        """Test workflow functionality with environment switching."""
-        os.chdir(self.temp_dir)
-
-        # Setup project
-        result = setup_command(guide=True, project_name="env_switch_test")
-        self.assertEqual(result, 0)
-
-        # Read and modify config to switch to production environment
-        config_path = Path("ADRI/config.yaml")
-        with open(config_path, 'r') as f:
-            config = yaml.safe_load(f)
-
-        # Verify we can switch environment in config
-        self.assertEqual(config["adri"]["default_environment"], "development")
-
-        # Switch to production
-        config["adri"]["default_environment"] = "production"
-        with open(config_path, 'w') as f:
-            yaml.dump(config, f)
-
-        # Verify environment switch worked
-        with open(config_path, 'r') as f:
-            updated_config = yaml.safe_load(f)
-
-        self.assertEqual(updated_config["adri"]["default_environment"], "production")
-
-        # Test that paths still resolve correctly after environment switch
-        tutorial_path = _resolve_project_path("tutorials/invoice_processing/invoice_data.csv")
-        # Use normalized path for cross-platform compatibility (Windows uses backslashes)
-        normalized_path = str(tutorial_path).replace("\\", "/")
-        self.assertTrue("ADRI/tutorials" in normalized_path)
-
-    def test_workflow_error_handling_and_recovery(self):
-        """Test workflow error handling and recovery scenarios."""
-        os.chdir(self.temp_dir)
-
-        # Test 1: Missing configuration error handling
-        with patch('click.echo') as mock_echo:
-            result = generate_standard_command("nonexistent/data.csv", guide=True)
-            self.assertEqual(result, 1)  # Should fail gracefully
-
-            # Verify helpful error message
-            echo_calls = [call.args[0] for call in mock_echo.call_args_list]
-            error_output = ' '.join(echo_calls)
-            self.assertIn("Generation failed", error_output)
-
-        # Test 2: Recovery after setup
-        result = setup_command(guide=True, project_name="error_recovery_test")
-        self.assertEqual(result, 0)
-
-        # Now the same command should work
-        with patch('adri.cli.load_data') as mock_load_data:
-            mock_load_data.return_value = [{"test": "data"}]
-            result = generate_standard_command("tutorials/invoice_processing/invoice_data.csv", force=True)
-            self.assertEqual(result, 0)
-
-    def test_workflow_performance_under_load(self):
-        """Test workflow performance with multiple operations."""
-        os.chdir(self.temp_dir)
-
-        # Setup project
-        result = setup_command(guide=True, project_name="performance_test")
-        self.assertEqual(result, 0)
-
-        # Time multiple path resolution operations
-        start_time = time.time()
-
-        paths_to_test = [
-            "tutorials/invoice_processing/invoice_data.csv",
-            "dev/standards/test_standard.yaml",
-            "prod/assessments/test_report.json",
-            "tutorials/customer_service/customer_data.csv",
-            "dev/training-data/snapshot.csv",
-        ] * 10  # Test 50 path resolutions
-
-        for path in paths_to_test:
-            resolved_path = _resolve_project_path(path)
-            self.assertIsInstance(resolved_path, Path)
-            self.assertTrue(resolved_path.is_absolute())
-
-        end_time = time.time()
-        total_time = end_time - start_time
-
-        # Performance assertion - should handle 50 resolutions quickly
-        self.assertLess(total_time, 2.0, f"Path resolution too slow: {total_time:.4f}s for 50 operations")
-
-    def test_workflow_with_mock_projects(self):
-        """Test workflow integration with mock project fixtures."""
-        # Test with simple project fixture
-        with mock_project_context(MockProjectFixtures.simple_adri_project()) as project_root:
-            # Verify project was created correctly
-            self.assertTrue((project_root / "ADRI" / "config.yaml").exists())
-            self.assertTrue((project_root / "ADRI" / "tutorials" / "invoice_processing" / "invoice_data.csv").exists())
-
-            # Test path resolution works in mock project
-            tutorial_path = _resolve_project_path("tutorials/invoice_processing/invoice_data.csv")
-            expected_path = project_root / "ADRI" / "tutorials" / "invoice_processing" / "invoice_data.csv"
-            self.assertEqual(tutorial_path.resolve(), expected_path.resolve())
-
-        # Test with complex project fixture
-        with mock_project_context(MockProjectFixtures.complex_multi_directory_project(), "docs/src") as project_root:
-            # Test from subdirectory
-            root_from_subdir = _find_adri_project_root()
-            self.assertEqual(root_from_subdir.resolve(), project_root.resolve())
-
-            # Test multiple path types from subdirectory
-            tutorial_path = _resolve_project_path("tutorials/financial_analysis/market_data.csv")
-            dev_path = _resolve_project_path("dev/standards/invoice_standard.yaml")
-
-            # Use normalized paths for cross-platform compatibility (Windows uses backslashes)
-            tutorial_normalized = str(tutorial_path.resolve()).replace("\\", "/")
-            dev_normalized = str(dev_path.resolve()).replace("\\", "/")
-            self.assertTrue("ADRI/tutorials" in tutorial_normalized)
-            self.assertTrue("ADRI/dev" in dev_normalized)
-
-
-class TestWorkflowIntegration(unittest.TestCase):
-    """End-to-end workflow testing across environments."""
-
-    def setUp(self):
-        """Set up test fixtures."""
-        self.temp_dir = tempfile.mkdtemp()
-        self.original_cwd = os.getcwd()
-        os.chdir(self.temp_dir)
-
-    def tearDown(self):
-        """Clean up test fixtures."""
-        os.chdir(self.original_cwd)
-        shutil.rmtree(self.temp_dir)
-
-    def test_complete_tutorial_workflow_end_to_end(self):
-        """Test the complete tutorial workflow from start to finish."""
-        # Step 1: Initialize project
-        result = setup_command(guide=True, project_name="tutorial_e2e")
-        self.assertEqual(result, 0)
-
-        # Verify tutorial files were created
-        invoice_data = Path("ADRI/tutorials/invoice_processing/invoice_data.csv")
-        test_data = Path("ADRI/tutorials/invoice_processing/test_invoice_data.csv")
-        self.assertTrue(invoice_data.exists())
-        self.assertTrue(test_data.exists())
-
-        # Verify file contents are correct
-        with open(invoice_data, 'r') as f:
-            content = f.read()
-            self.assertIn("invoice_id,customer_id,amount", content)
-            self.assertIn("INV-001,CUST-101,1250.00", content)
-
-        # Step 2: Generate standard with lineage tracking
-        with patch('adri.cli.load_data') as mock_load_data:
-            mock_load_data.return_value = [
-                {"invoice_id": "INV-001", "customer_id": "CUST-101", "amount": 1250.00}
-            ]
-
-            result = generate_standard_command("tutorials/invoice_processing/invoice_data.csv", guide=True, force=True)
-            self.assertEqual(result, 0)
-
-        # Verify standard was created with all sections
-        standard_path = Path("ADRI/dev/standards/invoice_data_ADRI_standard.yaml")
-        self.assertTrue(standard_path.exists())
-
-        with open(standard_path, 'r') as f:
-            standard = yaml.safe_load(f)
-
-        required_sections = ["training_data_lineage", "standards", "record_identification", "requirements", "metadata"]
-        for section in required_sections:
-            self.assertIn(section, standard)
-
-        # Verify training data snapshot was created
-        training_data_dir = Path("ADRI/dev/training-data")
-        snapshot_files = list(training_data_dir.glob("invoice_data_*.csv"))
-        self.assertEqual(len(snapshot_files), 1)
-
-        # Step 3: Run assessment
-        with patch('adri.cli.load_data') as mock_load_data, \
-             patch('adri.cli.load_standard') as mock_load_standard, \
-             patch('adri.cli.DataQualityAssessor') as mock_assessor_class:
-
-            # Setup mocks
-            mock_load_data.return_value = [{"invoice_id": "INV-101", "amount": 1350.00}]
-            mock_load_standard.return_value = standard
-
-            mock_result = Mock()
-            mock_result.overall_score = 88.5
-            mock_result.passed = True
-            mock_result.to_standard_dict.return_value = {"score": 88.5}
-
-            mock_assessor = Mock()
-            mock_assessor.assess.return_value = mock_result
-            mock_assessor.audit_logger = None
-            mock_assessor_class.return_value = mock_assessor
-
-            result = assess_command(
-                "tutorials/invoice_processing/test_invoice_data.csv",
-                "dev/standards/invoice_data_ADRI_standard.yaml",
-                guide=True
-            )
-            self.assertEqual(result, 0)
-
-    def test_workflow_cross_directory_consistency(self):
-        """Test that workflow produces consistent results from different directories."""
-        # Setup project
-        result = setup_command(guide=True, project_name="consistency_test")
-        self.assertEqual(result, 0)
-
-        project_root = Path.cwd()
-
-        # Test directories to run commands from
-        test_directories = [
-            "docs",
-            "src",
-            "tests",
-            "scripts",
-        ]
-
-        for test_dir in test_directories:
-            with self.subTest(directory=test_dir):
-                # Create test directory
-                test_path = project_root / test_dir
-                test_path.mkdir(exist_ok=True)
-                os.chdir(test_path)
-
-                # Test that path resolution is consistent
-                tutorial_path = _resolve_project_path("tutorials/invoice_processing/invoice_data.csv")
-                expected_path = project_root / "ADRI" / "tutorials" / "invoice_processing" / "invoice_data.csv"
-
-                self.assertEqual(tutorial_path, expected_path)
-
-                # Return to project root for next test
-                os.chdir(project_root)
-
-    def test_workflow_with_environment_documentation_integration(self):
-        """Test workflow integration with environment documentation features."""
-        result = setup_command(guide=True, project_name="env_doc_test")
-        self.assertEqual(result, 0)
-
-        # Verify config was created (focus on structure rather than specific documentation)
-        config_path = Path("ADRI/config.yaml")
-        with open(config_path, 'r') as f:
-            config_content = f.read()
-
-        # Check that basic environment structure exists
-        self.assertIn("environments:", config_content)
-        self.assertIn("development:", config_content)
-        self.assertIn("production:", config_content)
-
-        # Test show-config integration
-        with patch('click.echo') as mock_echo:
-            result = show_config_command()
-            self.assertEqual(result, 0)
-
-            # Handle mock call structure more safely
-            echo_calls = []
-            for call in mock_echo.call_args_list:
-                if call.args:
-                    echo_calls.append(call.args[0])
-                elif len(call) > 0 and hasattr(call[0], '__iter__'):
-                    echo_calls.extend(str(arg) for arg in call[0])
-
-            config_output = ' '.join(echo_calls)
-
-            # Should display both environments
-            self.assertIn("Development Environment:", config_output)
-            self.assertIn("Production Environment:", config_output)
-
-    def test_workflow_artifact_creation_and_validation(self):
-        """Test that workflow creates all expected artifacts correctly."""
-        result = setup_command(guide=True, project_name="artifact_test")
-        self.assertEqual(result, 0)
-
-        # Expected directory structure
-        expected_directories = [
-            "ADRI/tutorials/invoice_processing",
-            "ADRI/dev/standards",
-            "ADRI/dev/assessments",
-            "ADRI/dev/training-data",
-            "ADRI/dev/audit-logs",
-            "ADRI/prod/standards",
-            "ADRI/prod/assessments",
-            "ADRI/prod/training-data",
-            "ADRI/prod/audit-logs",
-        ]
-
-        for directory in expected_directories:
-            self.assertTrue(Path(directory).exists(), f"Directory should exist: {directory}")
-
-        # Expected configuration files
-        expected_files = [
-            "ADRI/config.yaml",
-            "ADRI/tutorials/invoice_processing/invoice_data.csv",
-            "ADRI/tutorials/invoice_processing/test_invoice_data.csv",
-        ]
-
-        for file_path in expected_files:
-            self.assertTrue(Path(file_path).exists(), f"File should exist: {file_path}")
-
-    def test_workflow_backward_compatibility(self):
-        """Test that enhanced workflow maintains backward compatibility."""
-        result = setup_command(guide=True, project_name="backward_compat_test")
-        self.assertEqual(result, 0)
-
-        # Test that old-style paths still work (with ADRI/ prefix)
-        with patch('adri.cli.load_data') as mock_load_data:
-            mock_load_data.return_value = [{"test": "data"}]
-
-            # Use old-style path with ADRI/ prefix
-            result = generate_standard_command("ADRI/tutorials/invoice_processing/invoice_data.csv", force=True)
-            self.assertEqual(result, 0)
-
-        # Verify standard was created correctly
-        standard_path = Path("ADRI/dev/standards/invoice_data_ADRI_standard.yaml")
-        self.assertTrue(standard_path.exists())
-
-
-class TestACTWorkflowValidation(unittest.TestCase):
-    """ACT-based GitHub workflow validation."""
-
-    def setUp(self):
-        """Set up test fixtures."""
-        self.temp_dir = tempfile.mkdtemp()
-        self.original_cwd = os.getcwd()
-
-    def tearDown(self):
-        """Clean up test fixtures."""
-        os.chdir(self.original_cwd)
-        shutil.rmtree(self.temp_dir)
-
-    def test_test_validation_workflow_structure(self):
-        """Test that test-validation.yml workflow is properly structured."""
-        # Check if the test validation workflow exists
-        workflow_path = Path(".github/workflows/test-validation.yml")
-
-        # If running from test environment, adjust path
-        if not workflow_path.exists():
-            workflow_path = Path(self.original_cwd) / ".github/workflows/test-validation.yml"
-
-        if workflow_path.exists():
-            with open(workflow_path, 'r') as f:
-                workflow = yaml.safe_load(f)
-
-            # Verify workflow structure
-            self.assertIn("jobs", workflow)
-            jobs = workflow["jobs"]
-
-            # Check for expected jobs
-            expected_jobs = [
-                "path-resolution-validation",
-                "enhanced-test-suite",
-                "workflow-compatibility",
-                "environment-documentation-check",
-                "integration-workflow-test",
-            ]
-
-            for job_name in expected_jobs:
-                self.assertIn(job_name, jobs, f"Workflow missing job: {job_name}")
-
-            # Verify job dependencies
-            integration_job = jobs.get("integration-workflow-test", {})
-            if "needs" in integration_job:
-                needs = integration_job["needs"]
-                self.assertIn("path-resolution-validation", needs)
-                self.assertIn("enhanced-test-suite", needs)
-
-    def test_act_configuration_validation(self):
-        """Test that ACT configuration is properly set up."""
-        actrc_path = Path(".actrc")
-
-        # If running from test environment, adjust path
-        if not actrc_path.exists():
-            actrc_path = Path(self.original_cwd) / ".actrc"
-
-        if actrc_path.exists():
-            with open(actrc_path, 'r') as f:
-                actrc_content = f.read()
-
-            # Check for essential ACT configurations
-            required_configs = [
-                "ubuntu-latest=catthehacker/ubuntu:act-latest",
-                "--container-architecture linux/amd64",
-                "--env CI=true",
-                "--env GITHUB_ACTIONS=true",
-            ]
-
-            for config in required_configs:
-                self.assertIn(config, actrc_content, f"ACT config missing: {config}")
-
-    def test_workflow_timeout_configurations(self):
-        """Test that workflows have appropriate timeout configurations."""
-        workflow_path = Path(".github/workflows/test-validation.yml")
-
-        if not workflow_path.exists():
-            workflow_path = Path(self.original_cwd) / ".github/workflows/test-validation.yml"
-
-        if workflow_path.exists():
-            with open(workflow_path, 'r') as f:
-                workflow = yaml.safe_load(f)
-
-            jobs = workflow.get("jobs", {})
-
-            for job_name, job_config in jobs.items():
-                with self.subTest(job=job_name):
-                    # Each job should have a timeout configured
-                    if "timeout-minutes" in job_config:
-                        timeout = job_config["timeout-minutes"]
-                        self.assertIsInstance(timeout, int)
-                        self.assertLessEqual(timeout, 15, f"Job {job_name} timeout too long: {timeout} minutes")
-
-
-if __name__ == '__main__':
-    unittest.main()
